+++ src/realPSet.cpp
@@ -605,8 +605,7 @@ void RealPSet::Prop()
     iter = m_posses.begin();
     j = 0;
     int size = m_posses.size();
-    //BAM par
-#pragma omp for schedule(static) 
+ #pragma omp for schedule(static) 
     for (int i = 0; i < size; ++i) {
       if (j > i) {
 	cout << "uhoh\n";
@@ -690,7 +689,6 @@ void RealPSet::Cull(Phase phase)
     iter = m_posses.begin();
     j = 0;
     int size = m_posses.size();
-    //BAM par
 #pragma omp for schedule(static) 
     for (int i = 0; i < size; ++i) {
       if (i < j) {
@@ -803,7 +801,6 @@ bool RealPSet::TakeIter(const TransMap &transMap,
     iter = m_posses.begin();
     j = 0;
     int size = m_posses.size();
-    //BAM par
 #pragma omp for schedule(static) 
     for (int i = 0; i < size; ++i) {
       while (j < i) {
@@ -982,16 +979,16 @@ void RealPSet::CombineAndRemoveTunnels()
       m_inTuns.erase(m_inTuns.begin()+i);
       for (unsigned int j = 0; j < m_leftInMap.size(); ++j) {
 	int val = m_leftInMap[j];
-	if (val == i)
+	if (val == (int)i)
 	  m_leftInMap[j] = -1;
-	if (val > i)
+	if (val > (int)i)
 	  m_leftInMap[j] = val-1;
       }
       for (unsigned int j = 0; j < m_rightInMap.size(); ++j) {
 	int val = m_rightInMap[j];
-	if (val == i)
+	if (val == (int)i)
 	  m_rightInMap[j] = -1;
-	else if (val > i)
+	else if (val > (int)i)
 	  m_rightInMap[j] = val-1;
       }
       --i;
@@ -1036,17 +1033,17 @@ void RealPSet::CombineAndRemoveTunnels()
 	
 	for(unsigned int j = 0; j < m_leftOutMap.size(); ++j) {
 	  int val = m_leftOutMap[j];
-	  if (val == i)
+	  if (val == (int)i)
 	    m_leftOutMap[j] = -1;
-	  else if (val > i)
+	  else if (val > (int)i)
 	    m_leftOutMap[j] = val-1;
 	}
 
 	for(unsigned int j = 0; j < m_rightOutMap.size(); ++j) {
 	  int val = m_rightOutMap[j];
-	  if (val == i)
+	  if (val == (int)i)
 	    m_rightOutMap[j] = -1;
-	  else if (val > i)
+	  else if (val > (int)i)
 	    m_rightOutMap[j] = val-1;
 	}
 
@@ -1116,7 +1113,6 @@ void RealPSet::Simplify(const TransMap &simplifiers, bool recursive)
     iter = m_posses.begin();
     j = 0;
     int size = m_posses.size();
-    //BAM par
 #pragma omp for schedule(static) 
     for (int i = 0; i < size; ++i) {
       while (j < i) {
@@ -1187,7 +1183,6 @@ bool RealPSet::MergePosses(const TransMap &simplifiers, CullFunction cullFunc)
   
   if (numPosses > 1) {
     PossMMap mmap;
-    //BAM par
     PossMMapIter iter;
     int j = 0;
 #pragma omp parallel private(j,iter)
@@ -1195,7 +1190,6 @@ bool RealPSet::MergePosses(const TransMap &simplifiers, CullFunction cullFunc)
       iter = m_posses.begin();
 	j = 0;
 	int size = m_posses.size();
-	//BAM par
 #pragma omp for schedule(static) 
       for (int i = 0; i < size; ++i) {
 	while (j < i) {
@@ -1536,7 +1530,7 @@ void RealPSet::InlinePoss(Poss *inliningPoss, PossMMap &newPosses)
     iter = newPoss->m_outTuns.begin();
     for(; iter != newPoss->m_outTuns.end(); ++iter) {
       Node *node = *iter;
-      for(int i = 0; i < node->m_inputs.size(); ++i) {
+      for(unsigned int i = 0; i < node->m_inputs.size(); ++i) {
 	Node *in = node->Input(i);
 	if (tunnelNumMap.find(in) != tunnelNumMap.end())
 	  throw;
@@ -1548,7 +1542,7 @@ void RealPSet::InlinePoss(Poss *inliningPoss, PossMMap &newPosses)
     iter = newPoss->m_possNodes.begin();
     for(; iter != newPoss->m_possNodes.end(); ++iter) {
       Node *node = *iter;
-      for(int i = 0; i < node->m_children.size(); ++i) {
+      for(unsigned int i = 0; i < node->m_children.size(); ++i) {
 	Node *child = node->Child(i);
 	if (node->m_poss != child->m_poss) {
 	  if (node->IsTunnel() && child->IsTunnel()) {
